Deleted Package and Tried to implement player controller with rigidbody
Deleted Standard Asset package because it was not needed.

Tried to implement the player controller from the Candy Mayhem game that used a rigidbody. I am going to need to change it a lot to get it to work with a capsule collider and FPS mouse look script.

Maybe try Character Controller route?