Camera_movement_part_2
Camera now moves to the -/+ target location. Instead of just adding the target angle to the start angle (which would result in a start angle of 20 being added to the target angle of  80 to equal 100 instead of moving to 80), it moves to the target angle by finding the difference between the target angle and the start angle.

Went through and commented everything for documentation.