Location-Point-Manager-Checks-for-1-trip
The LocationPointManager checks to see if the start location has been reached and the end location has been reached, then it ends its search.
todo:
need to make the player go to the stop location then back to the start location to end the game. Not just go to the stop location.